Added functionality to save and load files
Added the functions for these two actions
but no option to do them in the program, this
will come at a later date. Can now open files with inputting
the name of the file as the first argument like:

djinni myfile.txt

This would open the file 'myfile.txt' on Djinni.